requirements.txt
huggingface_hub
openai
langchain
langchain-community
streamlit
python-dotenv
